Fix infinite loop during unit testing
In unit testing measuring dimensions is impossible, so toast height
evaluates to `0`, which causes Toaster to endlessly recompute the height
and update the store. Instead we're being more specific and checking
whether height is `undefined`, and memoizing the ref so that it isn't
being called on every render. This also required other tweaks like
properly memoizing handlers like `updateHeight` so that they can be used as hook
dependencies. (We only needed `updateHeight` to be memoized, though.)